service: move framework files to awsconfig (#701)
