+++ service/awsconfig/error.go
@@ -1,4 +1,4 @@
-package service
+package awsconfig
 
 import (
 	"github.com/giantswarm/microerror"
@@ -9,4 +9,11 @@ var invalidConfigError = microerror.New("invalid config")
 // IsInvalidConfig asserts invalidConfigError.
 func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
+}
+
+var invalidVersionError = microerror.New("invalid version")
+
+// IsInvalidVersion asserts invalidVersionError.
+func IsInvalidVersion(err error) bool {
+	return microerror.Cause(err) == invalidVersionError
 }
