+++ service/awsconfig/framework.go
@@ -1,4 +1,4 @@
-package service
+package awsconfig
 
 import (
 	"context"
