+++ service/awsconfig/framework_test.go
@@ -1,4 +1,4 @@
-package service
+package awsconfig
 
 import (
 	"reflect"
