+++ service/awsconfig/resource_router.go
@@ -1,4 +1,4 @@
-package service
+package awsconfig
 
 import (
 	"context"
