+++ service/awsconfig/resource_router_test.go
@@ -1,4 +1,4 @@
-package service
+package awsconfig
 
 import (
 	"context"
